feat: change structure of project
add axios error handler
