+++ src/app/App.tsx
@@ -1,17 +1,18 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
-import GlobalStyle from 'app/styles/GlobalStyle';
+import { GlobalStyle } from 'app/styles';
 import Header from 'shared/ui/header/Header';
-import AddItem from 'pages/additem/AddItemPage';
-import Items from 'pages/items/ItemsPage';
-import NotFound from 'shared/ui/notfound/NotFound';
+import Home from 'pages/HomePage';
+import AddItem from 'pages/AddItemPage';
+import Items from 'pages/ItemsListPage';
+import NotFound from 'pages/NotFoundPage';
 
-function App() {
+export default function App() {
   return (
     <Router>
       <GlobalStyle />
       <Header />
       <Routes>
-        <Route path='/' element={<Items />} />
+        <Route path='/' element={<Home />} />
         <Route path='/items' element={<Items />} />
         <Route path='/additem' element={<AddItem />} />
         <Route path='*' element={<NotFound />} />
@@ -19,5 +20,3 @@ function App() {
     </Router>
   );
 }
-
-export default App;
