+++ src/app/styles/GlobalStyle.tsx
@@ -1,6 +1,6 @@
 import { createGlobalStyle } from 'styled-components';
 
-const GlobalStyle = createGlobalStyle` 
+export const GlobalStyle = createGlobalStyle` 
 @font-face {
     font-family: 'Pretendard-Regular';
     src: url('https://fastly.jsdelivr.net/gh/Project-Noonnu/noonfonts_2107@1.1/Pretendard-Regular.woff') format('woff');
@@ -38,4 +38,3 @@ a {
     text-decoration: none;
 }
 `;
-export default GlobalStyle;
