+++ src/features/itemlist/lib/config.ts
@@ -1,0 +1,2 @@
+// 환경변수 타입 정의
+export const BASE_URL: string | undefined = process.env.REACT_APP_API_URL;
