+++ src/features/itemlist/lib/index.ts
@@ -1,0 +1,3 @@
+export { BASE_URL } from './config';
+export type { GetItemsParams, Item, GetItemsResponse } from './api';
+export { getItems } from './getItems';
