+++ src/index.jsx
@@ -1,5 +1,10 @@
+import { StrictMode } from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './app/App';
+import App from 'app/App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(<App />);
+root.render(
+  <StrictMode>
+    <App />
+  </StrictMode>,
+);
