+++ src/pages/HomePage.tsx
@@ -1,0 +1,5 @@
+function HomePage() {
+  return <h2>임시 페이지</h2>;
+}
+
+export default HomePage;
