+++ src/shared/assets/panda-market-logo-mobile.png
