+++ src/shared/assets/panda-market-logo.png
