+++ src/shared/ui/NavigationLink.tsx
